+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `vanity import --scrape` flag - Scrape the GitHub profile page to capture private contributions that aren't available via the API
+
 ## [0.2.0] - 2026-01-31
 
 ### Added